+++ packages/server/src/internals/middlewares.ts
@@ -1,0 +1,19 @@
+import { ProcedureType } from '../router';
+
+export const middlewareMarker = Symbol('middlewareMarker');
+interface MiddlewareResult {
+  /**
+   * All middlewares should pass through their `next()`'s output.
+   * Requiring this marker makes sure that can't be forgotten at compile-time.
+   */
+  readonly marker: typeof middlewareMarker;
+  output: unknown;
+}
+
+export type MiddlewareFunction<TContext> = (opts: {
+  ctx: TContext;
+  type: ProcedureType;
+  path: string;
+  input: unknown;
+  next: () => Promise<MiddlewareResult>;
+}) => Promise<MiddlewareResult>;
